fix type issue

+++ ex2_hal/iris/equipment_handler/source/iris_spi.c
@@ -122,7 +122,7 @@ void iris_spi_get(uint16_t *rx_data, uint16_t data_length) {
  *   Number of ticks for non FreeRTOS delay
  *
  **/
-void iris_spi_delay(uint16_t ticks) {
+void iris_spi_delay(uint32_t ticks) {
     uint16_t i;
     for (i = 0; i < ticks; i++)
         ;
